+++ packages/client/ui/AlertDialog/AlertDialogContent.tsx
@@ -21,3 +21,4 @@ const AlertDialogContent = React.forwardRef<
   </AlertDialogPortal>
 ))
 AlertDialogContent.displayName = AlertDialogPrimitive.Content.displayName
+export {AlertDialogContent}